requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pip==22.3.1
py-cord==2.4.1
python-dotenv==1.0.0
setuptools==65.5.1
wheel==0.38.4
yarl==1.9.2
